+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/ConfigurationTest.php
@@ -52,6 +52,34 @@ class ConfigurationTest extends TestCase
         $this->assertEquals(self::getBundleDefaultConfig(), $config);
     }
 
+    public function testRateLimiterBuilderIsNotReadAsALimiterName()
+    {
+        $processor = new Processor();
+        $config = $processor->processConfiguration(new Configuration(true), [[
+            // no "limiters" key: the shorthand below would otherwise read "builder" as a limiter name
+            'rate_limiter' => [
+                'builder' => ['cache_pool' => 'my.pool'],
+            ],
+        ]]);
+
+        $this->assertSame([], $config['rate_limiter']['limiters']);
+        $this->assertSame('my.pool', $config['rate_limiter']['builder']['cache_pool']);
+    }
+
+    public function testRateLimiterBuilderCanBeConfiguredAlongsideLimiters()
+    {
+        $processor = new Processor();
+        $config = $processor->processConfiguration(new Configuration(true), [[
+            'rate_limiter' => [
+                'limiters' => ['foo' => ['policy' => 'fixed_window', 'limit' => 5, 'interval' => '1 minute']],
+                'builder' => ['cache_pool' => 'my.pool'],
+            ],
+        ]]);
+
+        $this->assertSame(['foo'], array_keys($config['rate_limiter']['limiters']));
+        $this->assertSame('my.pool', $config['rate_limiter']['builder']['cache_pool']);
+    }
+
     public function testTranslatorProviderDomainsCanBeKeyed()
     {
         $processor = new Processor();
@@ -1334,6 +1362,11 @@ class ConfigurationTest extends TestCase
             'rate_limiter' => [
                 'enabled' => !class_exists(FullStack::class) && class_exists(TokenBucketLimiter::class),
                 'limiters' => [],
+                'builder' => [
+                    'lock_factory' => 'auto',
+                    'cache_pool' => 'cache.rate_limiter',
+                    'storage_service' => null,
+                ],
             ],
             'uid' => [
                 'enabled' => !class_exists(FullStack::class) && class_exists(UuidFactory::class),