+++ src/Symfony/Component/RateLimiter/CHANGELOG.md
@@ -5,6 +5,8 @@ CHANGELOG
 ---
 
  * Add `RateLimitExceededEvent`
+ * Add `RateLimiterBuilder`
+ * Allow `\DateInterval` for the `interval` and `rate.interval` options of `RateLimiterFactory`
 
 8.1
 ---